LostarkDiscordBot/requirements.txt
discord.py
asyncio
requests
beautifulsoup4